==> R/33_fish-spp-distributions.R <==

# Plot fish species occurance and rel. abundance along gradient

# Prep -----

source(here::here("R", "00_prep.R"))
fish_meta <- read_csv(here("data", "metadata_fishes.csv"))

# some labels
fam_levels <- c("Catostomidae","Centrarchidae","Clupeidae",
                "Cyprinidae","Fundulidae","Gasterosteridae",
                "Ictaluridae","Percidae","Salmonidae","Sciaenidae")
origin_levels <- c("Native","NonNative")

## Data ------------------

fish_spp_abundnace <- read_csv(here("out", "r1_fish_spp_counts.csv"))

den <- fish_spp_abundnace %>% 
  filter(stream_name != "Horse") %>% 
  group_by(site_id, taxon_code) %>%
  summarise(density = mean(density))

data <- data_field %>% 
  group_by(site_id, taxon_code) %>%
  summarise(count = sum(count)) %>% 
  left_join(gradient, by = "site_id") %>% 
  left_join(fish_meta, by = "taxon_code") %>% 
  mutate(site_group = factor(site_group, levels = c("Upstream", "Midstream", "Downstream")), 
         family = factor(family, levels = fam_levels), 
         origin = factor(origin, levels = origin_levels)) %>% 
  left_join(den, by = c("site_id", "taxon_code"))
data

# Plotting ----------------------------------------------------------------

# Plot
spp_dist <- 
  data %>% 
  group_by(common_name) %>% 
  ggplot(aes(x = fct_reorder(common_name, PC1, .desc = TRUE), y = PC1)) + 
  coord_flip(clip = "off") +
  geom_point(aes(fill = trophic_group, shape = origin, size = density), color = "black") +
  scale_shape_manual(values = c(21, 24), name = "Origin") +
  labs(y = "Longitunal Gradient (PC1)",
       x = "Fish species", size = "Density",
       fill = "Trophic Group") +
  scale_fill_brewer(palette = "Set1") +
  scale_y_continuous(expand=c(0,0), limits=c(-0.25,8), breaks = seq(0,8,2)) +
  guides(fill = guide_legend(override.aes=list(shape=21))) +
  theme(legend.key.size = unit(0.5, "cm"), 
        legend.title = element_text(face = "plain"), 
        panel.grid.major = element_line(color = "grey", size = 0.25))
spp_dist


# ggsave(filename = "figs/spp-distribution.png", plot = last_plot(), 
#        units = "in", width = 11, height = 6)


#===============================================================================
# Panel Plot with other data
#===============================================================================


top_row <- plot_grid(
  p.richness + theme(plot.margin=unit(c(1,1,1,1),"mm")), 
  p.desnity + theme(plot.margin=unit(c(1,1,1,1),"mm"),
                    legend.position = "none"), 
  p.shannon + theme(plot.margin=unit(c(1,1,1,1),"mm"),
                    legend.position = "none"),
  labels = c("A", "B", "C"), ncol = 3, nrow = 1,
  align = "vh"
  )

panel <- plot_grid(top_row, spp_dist, labels = c('', 'D'), ncol = 1, rel_heights = c(0.7, 1))
panel

# Save it
path <- here::here("out", "r1_fish_comm_panel")
ggsave(glue::glue("{path}.pdf"), plot = panel, 
       width = 11, height = 8, device = cairo_pdf)
pdftools::pdf_convert(pdf = glue::glue("{path}.pdf"),
                      filenames = glue::glue("{path}.png"),
                      format = "png", dpi = 300)
