==> R/44_rkin-estimation.Rmd <==
---
title: "rKIN niche estimation"
output: html_notebook
editor_options: 
  chunk_output_type: console
---

```{r set up}
source(here::here("R", "00_prep.R"))
source(here("R", "40_prep-sia-data.R"))
```

```{r load data}

rich <- read_csv(here("out", "r1_fish_structure_summary.csv")) 
gradient <- read_csv(here("out", "r1_PCA_results.csv"))
fish_meta <- read_csv(here("data", "metadata_fishes.csv")) 

# fish_native <- sia_fish %>% 
#   left_join(fish_meta, by = "taxon_code") %>% 
#   filter(origin == "Native")
# 
# fish_nonnative <- sia_fish %>% 
#   left_join(fish_meta, by = "taxon_code") %>% 
#   filter(origin == "NonNative")

# Prepare fish data
fish_df <- sia_fish %>% 
  filter(sample_year %in% c(2016,2017)) %>%
  # filter(! (yr_site %in% c("2014_LR04","2014_LR05"))) %>% 
  # filter(!(sample_year==2015 & site_id == "LR02")) %>% 
  mutate_if(is.character, as.factor)

# Pred fish data
invert_df <- sia_inverts %>% 
  filter(sample_year %in% c(2016)) %>% 
  mutate_if(is.character, as.factor)


```

## Community-level
1. Fish community
```{r fish community}


# Calculate community niche breadth
kin_fish <- fish_df %>% 
  as.data.frame() %>% 
  estKIN(data = ., x = "d13Cc", y = "d15N", group = "yr_site",
         levels = 95, smallSamp = FALSE)

kin_fish_tbl <- kin_fish %>% 
  getArea() %>% 
  as_tibble() %>% 
  separate(Group, c("sample_year", "site_id"), sep = "_", remove = FALSE) %>% 
  left_join(gradient, by = "site_id") %>% 
  select(-site_id) %>% 
  rename(yr_site = Group, KUD = ShapeArea)

# Carbon range
fish_CR <- fish_df %>% 
  group_by(yr_site) %>% 
  summarise(CR = max(d13Cc)-min(d13Cc), .groups='drop')

# Nitrogen range
fish_NR <- fish_df %>% 
  group_by(yr_site) %>% 
  summarise(NR = max(d15N)-min(d15N), .groups='drop')

# Nearest neighbor
fish_MNND <- fish_df %>% 
  select(yr_site, d15N, d13C) %>% 
  group_by(yr_site) %>% 
  nest() %>% 
  mutate(NND = map(data, spatstat.geom::nndist)) %>% 
  mutate(MNND = map_dbl(NND, mean)) %>% 
  mutate(sdMNND = map_dbl(NND, sd)) %>% 
  select(-data, -NND)


# TCombine metrics
fish_comm_metrics <- kin_fish_tbl %>% 
  left_join(fish_CR, by = "yr_site") %>% 
  left_join(fish_NR, by = "yr_site") %>% 
  left_join(fish_MNND, by = "yr_site") %>% 
  select(stream_name, yr_site, PC1, PC2, KUD, CR, NR, MNND, sdMNND) %>% 
  separate(yr_site, into = c("sample_year","site_id"), remove = FALSE) %>% 
  #mutate(compartment = "fish") %>% 
  mutate_if(is.character, as.factor) %>% 
  arrange(sample_year, site_id)
fish_comm_metrics %>% print(n = Inf)


# Add richeness and abundance
rich <- rich %>% 
  # filter(sample_year %in% c(2015,2016,2017)) %>% 
  select(sample_year, site_id, richness, abund, density) %>% 
  mutate(sample_year = as_factor(sample_year)) %>% 
  mutate(richness = if_else(sample_year=="2016"&site_id=="LR07", 12, richness))

fish_comm_metrics <- fish_comm_metrics %>%
  left_join(rich, by = c("sample_year", "site_id"))%>% 
  # left_join(seac, by = "yr_site")%>% 
  mutate_if(is.character, as.factor)

```

2. Invert commuimty
```{r invert community}

# Calculate community niche breadth
kin_invert <- invert_df %>% 
  as.data.frame() %>% 
  estKIN(data = ., x = "d13Cc", y = "d15N", group = "yr_site", 
         levels = 95, smallSamp = FALSE)

kin_invert_tbl <- kin_invert %>% 
  getArea() %>% 
  as_tibble() %>% 
  separate(Group, c("sample_year", "site_id"), sep = "_", remove = FALSE) %>% 
  left_join(gradient, by = "site_id") %>% 
  select(-site_id) %>% 
  rename(yr_site = Group, KUD = ShapeArea)

# Carbon range
invert_CR <- invert_df %>% 
  group_by(yr_site) %>% 
  summarise(CR = max(d13Cc)-min(d13Cc), .groups='drop')

# Nitrogen range
invert_NR <- invert_df %>% 
  group_by(yr_site) %>% 
  summarise(NR = max(d15N)-min(d15N), .groups='drop')

# Tbl for modeling
invert_comm_metrics <- kin_invert_tbl %>% 
  left_join(invert_CR, by = "yr_site") %>% 
  left_join(invert_NR, by = "yr_site") %>% 
  select(stream_name, yr_site, PC1, PC2, KUD, CR, NR) %>% 
  separate(yr_site, into = c("sample_year","site_id"), remove = FALSE) %>% 
  #mutate(compartment = "invert") %>% 
  mutate_if(is.character, as.factor) %>% 
  arrange(sample_year, site_id)
invert_comm_metrics %>% print(n = Inf)

```

3. Whole community (fish + inverts)
```{r fish community}
# Pred fish data
consumers_df <- sia_consumers %>% 
  filter(sample_year %in% c(2016)) %>%
  mutate_if(is.character, as.factor)

# Calculate community niche breadth
kin_comsumers <- consumers_df %>% 
  as.data.frame() %>% 
  estKIN(data = ., x = "d13Cc", y = "d15N", group = "yr_site",  h = "hnm",
         levels = 95, smallSamp = FALSE)

kin_consumers_tbl <- kin_comsumers %>% 
  getArea() %>% 
  as_tibble() %>% 
  separate(Group, c("sample_year", "site_id"), sep = "_", remove = FALSE) %>% 
  left_join(gradient, by = "site_id") %>% 
  select(-site_id) %>% 
  rename(yr_site = Group, KUD = ShapeArea)

# Carbon range
consumer_CR <- consumers_df %>% 
  group_by(yr_site) %>% 
  summarise(CR = max(d13Cc)-min(d13Cc), .groups='drop')

# Nitrogen range
consumer_NR <- consumers_df %>% 
  group_by(yr_site) %>% 
  summarise(NR = max(d15N)-min(d15N), .groups='drop')


# Tbl for modeling
consumer_comm_metrics <- kin_consumers_tbl %>% 
  left_join(consumer_CR, by = "yr_site") %>% 
  left_join(consumer_NR, by = "yr_site") %>% 
  select(stream_name, yr_site, PC1, PC2, KUD, CR, NR) %>% 
  separate(yr_site, into = c("sample_year","site_id"), remove = FALSE) %>% 
  # mutate(KUD = if_else(site_id %in% c("MB03", "SW03"), KUD+5, KUD)) |>
  # mutate(KUD = if_else(site_id %in% c("LR05"), KUD+10, KUD)) |>
  mutate_if(is.character, as.factor) %>% 
  arrange(sample_year, site_id) 

consumer_comm_metrics %>% print(n = Inf)
```



## Species-level fish 

```{r nest data}
fish_df_spp <- fish_df %>% 
  select(-sample_hitch, - compartment, -resource) %>% 
  mutate(taxon_code = case_when(
    taxon_code == "BNT" & length_mm < 100 ~ "BNTyoy",
    TRUE ~ as.character(taxon_code))) %>%
  mutate(taxon_code = case_when(
    taxon_code == "CKC" & length_mm < 75 ~ "CKCyoy",
    TRUE ~ as.character(taxon_code))) %>%
  mutate(taxon_code = case_when(
    taxon_code == "SMB" & length_mm < 75 ~ "SMByoy",
    TRUE ~ as.character(taxon_code))) %>%
  mutate_if(is.character, as.factor)  %>% 
  filter(! (sample_year == 2016 & site_id == "LR06" & taxon_code == "CSH")) %>% 
  # Remove species with fewer than 3 observations per site
  group_by(sample_year, site_id, taxon_code) %>% 
  filter(n() >= 3) %>% 
  ungroup() 

nested <- fish_df_spp %>% 
  group_by(sample_year, site_id) %>% 
  nest() 
nested

```


1. Calculate KUDs for each species in a site
```{r calc KUDs}
nested_kuds <- nested %>% 
  mutate(data_df = map(data, as.data.frame)) %>% 
  mutate(kud_obj = map(
    data_df, 
    ~estKIN(data = ., x = "d13Cc", y = "d15N", group = "taxon_code", 
            levels = 95, smallSamp = TRUE)))
nested_kuds  
```

2. Get niche area for each species
```{r calc spp-specific KUDs} 
spp_area <- nested_kuds %>% 
  mutate(area = map(kud_obj, getArea)) %>% 
  mutate(area = map(area, as_tibble)) %>% 
  unnest(area) %>% 
  ungroup() %>% 
  rename(species = Group, KUD95 = ShapeArea) %>% 
  select(sample_year, site_id, species, KUD95) %>% 
  left_join(gradient, by = "site_id")
spp_area
```

3. Get area of each species' niche occupied by other species' niches

```{r}

fish_df_spp %>% 
  filter(yr_site=="2016_LR01") %>% 
  mutate(taxon_code = ifelse(taxon_code=="BKT","BKT","other")) %>% 
  group_by(sample_year, site_id) %>% 
  nest() %>% 
  mutate(data_df = map(data, as.data.frame)) %>% 
  mutate(kud_obj = map(data_df, 
    ~estKIN(data = ., x = "d13Cc", y = "d15N", group = "taxon_code", 
            levels = 95, smallSamp = TRUE))) %>% 
  mutate(ovlp = map(kud_obj, calcOverlap)) %>% 
  unnest(ovlp) %>% 
  ungroup() %>% 
  filter(str_detect(OverlapID, "BKT")) %>% 
  select(OverlapID, starts_with("other"))

get_spp_ovlp <- function(data, target, species){
  data %>% 
  filter(yr_site==target) %>% 
  mutate(taxon_code = ifelse(taxon_code==species,species,"other")) %>% 
  group_by(sample_year, site_id) %>% 
  nest() %>% 
  mutate(data_df = map(data, as.data.frame)) %>% 
  mutate(kud_obj = map(data_df, 
    ~estKIN(data = ., x = "d13Cc", y = "d15N", group = "taxon_code", 
            levels = 95, smallSamp = TRUE))) %>% 
  mutate(ovlp = map(kud_obj, calcOverlap)) %>% 
  unnest(ovlp) %>% 
  ungroup() %>% 
  filter(str_detect(OverlapID, species)) %>% 
  select(OverlapID, starts_with("other"))
}

# test it
# fish_df_spp %>% filter(yr_site=="2016_LR00") %>% distinct(taxon_code)
# get_spp_ovlp(fish_df_spp, "2016_LR00", "BKT")
# get_spp_ovlp(fish_df_spp, "2016_LR00", "BNT")
# get_spp_ovlp(fish_df_spp, "2016_LR00", "LNS")
# 
# out <- list()
# sp <- c("BKT","BNT","LNS")
# for(i in sp) {
#   out[[i]] <- get_spp_ovlp(fish_df_spp, "2016_LR00", i)
# }
# out


target <- levels(fish_df_spp$yr_site)
sp <-  levels(fish_df_spp$taxon_code)

result <- list()

for(i in target) {
  
  result_nested <- list()
  
  for(j in sp){
    
    dat <- fish_df_spp %>% 
      filter(yr_site == i) %>%
      mutate(taxon_code = ifelse(taxon_code==j,j,"other")) %>% 
      as.data.frame()
    
    kud_obj <- estKIN(data = dat, x = "d13Cc", y = "d15N", group = "taxon_code", 
                levels = 95, smallSamp = TRUE)
    
    overlap <- calcOverlap(kud_obj)
    
    out <- overlap %>% 
      filter(!str_detect(OverlapID, "other")) %>% 
      select(starts_with("other")) %>% 
      pull()
    
    result_nested[[j]] <- out
  }
  
  result[[i]] <- result_nested
  
}

result
tmp <- as.data.frame(do.call(rbind, result))

spp_ovlp2 <- tmp %>% 
  rownames_to_column("yr_site") %>% 
  pivot_longer(-yr_site, values_to = "overlap") %>% 
  unnest(overlap) %>% 
  separate(yr_site, into = c("sample_year", "site_id"))

spp_ovlp2
```


```{r calc spp-specific overlap}
spp_ovlp <- nested_kuds %>% 
  mutate(ovlp = map(kud_obj, calcOverlap)) %>% 
  unnest(ovlp) %>% 
  ungroup() %>% 
  select(-sample_year, -site_id, -data, -data_df,-kud_obj, -OverlapID) %>% 
  mutate(mean_overlap = rowMeans(., na.rm = TRUE)) %>% 
  select(mean_overlap)  %>% 
  bind_cols(spp_area) 
spp_ovlp

rich <- rich %>% 
  filter(sample_year %in% c(2016,2017)) %>% 
  select(sample_year, site_id, richness, abund, density) %>% 
  mutate(sample_year = as.character(sample_year)) %>% 
  mutate(richness = if_else(sample_year=="2016"&site_id=="LR07", 12, richness))

# add spee_pvlp2 and rich
spp_ovlp3 <- spp_ovlp %>% 
  mutate(sample_year = as.character(sample_year)) %>% 
  left_join(spp_ovlp2, by = c("sample_year", "site_id", "species"="name")) %>% 
  mutate(sample_year = factor(sample_year)) %>% 
  left_join(rich, by = c("sample_year", "site_id")) %>% 
  relocate(overlap, .after = mean_overlap)
spp_ovlp3


# spp_ovlp3 %>% 
#   write_csv("data-derived/fish_overlap.csv")
```

```{r plot species KUDs vs PC1}
spp_ovlp3 %>% 
  group_by(species) %>% 
  # filter(! sample_year == 2015) %>% 
  #filter(site_id %in% c("LR06","LR07","LR08")) %>% 
  ggplot(aes(x = fct_reorder(species, PC1, .desc = TRUE), y = PC1)) + 
  coord_flip() +
  geom_point(aes(size = KUD95, fill = KUD95), shape = 21, color = "black") +
  labs(y = "Long. Gradient (PC1)",
       x = "Fish species", 
       fill = "KUD", size = "KUD") +
  scale_y_continuous(expand=c(0,0), limits=c(-0.25,9), breaks = seq(0,9,1)) +
  viridis::scale_fill_viridis(option = "E", discrete = FALSE) + 
  scale_size_continuous(breaks = seq(0,40,5)) + 
  guides(size = FALSE) + 
  theme(legend.key.size = unit(0.5, "cm"), 
        legend.title = element_text(face = "plain"), 
        panel.grid.major = element_line(color = "grey", size = 0.25)) + 
  lemon::facet_rep_wrap(vars(sample_year))
```

```{r plot overlap vs PC1 by species}
spp_ovlp %>% 
  group_by(species) %>% 
  ggplot(aes(x = fct_reorder(species, PC1, .desc = TRUE), y = PC1)) + 
  coord_flip() +
  geom_point(aes(size = mean_overlap, fill = mean_overlap), shape = 21, color = "black") +
  labs(y = "Long. Gradient (PC1)",
       x = "Fish species", 
       fill = "Overlap", size = "Overlap") +
  scale_y_continuous(expand=c(0,0), limits=c(-0.25,9), breaks = seq(0,9,1)) +
  scale_size_area() +
  viridis::scale_fill_viridis(option = "D", discrete = FALSE) + 
  guides(color = FALSE, fill = guide_legend(override.aes=list(shape=21))) + 
  theme(legend.key.size = unit(0.5, "cm"), 
        legend.title = element_text(face = "plain"), 
        panel.grid.major = element_line(color = "grey", size = 0.25)) + 
  lemon::facet_rep_wrap(vars(sample_year))
```

```{r plot overlap vs PC1}
spp_ovlp3 %>% 
  ggplot(aes(x = PC1, y = mean_overlap)) + 
  geom_point(aes(fill = mean_overlap), shape = 21, color = "black") +
  ggpubr::stat_cor(method = "pearson", label.x = 0.5, label.y = 1, vjust = 1) + 
  labs(x = "Long. Gradient (PC1)",
       y = "Overmap", 
       fill = "Overlap", size = "Overlap") +
  scale_x_continuous(expand=c(0,0), limits=c(-0.25,9), breaks = seq(0,9,1)) +
  viridis::scale_fill_viridis(option = "A", discrete = FALSE) + 
  guides(color = FALSE, fill = guide_legend(override.aes=list(shape=21))) + 
  theme(legend.key.size = unit(0.5, "cm"), 
        legend.title = element_text(face = "plain"), 
        panel.grid.major = element_line(color = "grey", size = 0.25))


spp_ovlp3 %>% 
  ggplot(aes(x = PC1, y = overlap)) + 
  geom_point(aes(fill = overlap), shape = 21, color = "black") +
  ggpubr::stat_cor(method = "pearson", label.x = 0.5, label.y = 1, vjust = 1) + 
  labs(x = "Long. Gradient (PC1)",
       y = "Overmap", 
       fill = "Overlap", size = "Overlap") +
  scale_x_continuous(expand=c(0,0), limits=c(-0.25,9), breaks = seq(0,9,1)) +
  viridis::scale_fill_viridis(option = "A", discrete = FALSE) + 
  guides(color = FALSE, fill = guide_legend(override.aes=list(shape=21))) + 
  theme(legend.key.size = unit(0.5, "cm"), 
        legend.title = element_text(face = "plain"), 
        panel.grid.major = element_line(color = "grey", size = 0.25))
```
